requirements.txt
git+git://github.com/nmercer/echo-python.git
#amazon-echo==0.5

beautifulsoup4==4.3.2
enum34==1.0.4
gnureadline==6.3.3
netifaces==0.10.4
protobuf==3.0.0a3
PyChromecast==0.6.4.1
python-nest==2.3.1
requests==2.6.0
six==1.9.0
zeroconf==0.17.1
